DP_CHANNEL_IDENTIFIER_CD
git-svn-id: https://skvalley.com/svn/repos/storeplatform/trunk/sac/storeplatform-sac@5237 01f5c84a-aead-4b29-b0f1-7c54dbeec05f

+++ src/main/java/com/skplanet/storeplatform/sac/display/feature/best/vo/BestApp.java
@@ -34,7 +34,7 @@ public class BestApp {
 	private String menuNesc;
 	private int prodAmt;
 	private String apkVerCd;
-	private int avgEvluScore;
+	private Double avgEvluScore;
 	private Integer dwldCnt;
 	private Integer paticpersCnt;
 	private String imgPath;
@@ -516,28 +516,11 @@ public class BestApp {
 		this.apkVerCd = apkVerCd;
 	}
 
-	/**
-	 * 
-	 * <pre>
-	 * 평점.
-	 * </pre>
-	 * 
-	 * @return int
-	 */
-	public int getAvgEvluScore() {
+	public Double getAvgEvluScore() {
 		return this.avgEvluScore;
 	}
 
-	/**
-	 * 
-	 * <pre>
-	 * 평점.
-	 * </pre>
-	 * 
-	 * @param avgEvluScore
-	 *            avgEvluScore
-	 */
-	public void setAvgEvluScore(int avgEvluScore) {
+	public void setAvgEvluScore(Double avgEvluScore) {
 		this.avgEvluScore = avgEvluScore;
 	}
